Adding company and user unit test fix

+++ database/migrations/2023_08_17_022243_create_e-contract_project_attachments_table.php
@@ -42,6 +42,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('e-contract_project_attachments');
+        if (DB::getDriverName() !== 'sqlite') {
+            Schema::dropIfExists('e-contract_project_attachments');
+        }
     }
 };
