+++ database/migrations/2023_09_08_112722_create_invoice_items_table.php
@@ -36,6 +36,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('invoice_items');
+        if (DB::getDriverName() !== 'sqlite') {
+            Schema::dropIfExists('invoice_items');
+        }
     }
 };
